LayoutTests/fast/dom/variadic-operations-length-expected.txt

Tests that the value of the 'length' property is correct for variadic operations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.body.before.length is 0
PASS document.body.after.length is 0
PASS document.body.replaceWith.length is 0
PASS document.body.prepend.length is 0
PASS document.body.append.length is 0
PASS document.body.classList.add.length is 0
PASS document.body.classList.remove.length is 0
PASS successfullyParsed is true

TEST COMPLETE
